Encapsulate model and navigation
The class `MouldModel` now encapsulated the character
and progress tracks. `MouldNavigation` provides event
based navigation methods, encapsulating the nav
controller. Both are exclusively passed to the top view
composables.

This provides an uniform API for the views and is a
first step to encapsulate the views to simplify and
homogenize their API and provide a top-level scaffold.